removed the page transition from home to bin to make it more responsive since loading two instances of Google Maps is actually quite expensive on the hardware.
removed access token testing.

updated getClosestBin to getClosestBins which matches the new API route, it now gets three bins instead of one.

the getClosestBins HTTP request has been limited to once every 1:20, so we're making less requests, but still polling for the closest bins as the user moves around. Need to test this to see if it's an appropriate solution.